Fetch initial movie

+++ recommendation-api/src/main/java/com/example/recommendationapi/dtos/MovieResponse.java
@@ -5,5 +5,5 @@ import java.util.Date;
 
 public record MovieResponse(int tmdbId, String title, String description,
                             Date releaseDate, Collection<String> genres, int runtime,
-                            int rating, String coverArtUrl, String trailerUrl) {
+                            double rating, String coverArtUrl, String trailerUrl) {
 }
